+++ src/main/java/com/example/ventas_bodega/repository/PlanRepository.java
@@ -3,7 +3,7 @@ package com.example.ventas_bodega.repository;
 import com.example.ventas_bodega.entity.PlanEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PlanRepository extends JpaRepository<PlanEntity, Integer> {
+public interface PlanRepository extends JpaRepository<PlanEntity, Long> {
